requirements.txt
python-telegram-bot==21.11.1
requests==2.32.3
python-dotenv==1.0.0 